+++ CHANGELOG.md
@@ -3,7 +3,11 @@ Changelog
 
 # 8.0.0
 
-- Bump bpmn-elements to v0.6.0 with refactored MessageEventDefinition, ReceiveTask, and some bug fixes from TallaInc
+## Breaking changes
+- Bump bpmn-elements to v0.6.0 with refactored MessageEventDefinition, ReceiveTask, and some bug fixes from @TallaInc
+
+## Additions
+- Recover takes optional options that completely overrides environment and passes it along to the recovered definitions
 
 # 7.1.0